+++ sarkit/xmlhelp/_core.py
@@ -185,13 +185,25 @@ class ElementWrapper(collections.abc.MutableMapping):
         self.typename = typename
         self.typedef = xsdhelper.xsdtypes[typename]
 
+    def _getchilddef(self, localname):
+        childdef = self.typedef.get_childdef_from_localname(localname)
+        if childdef is None:
+            raise KeyError(localname)
+        return childdef
+
+    def _getattribname(self, localname):
+        attribname = self.typedef.get_attribute_from_localname(
+            localname.removeprefix("@")
+        )
+        if attribname is None or self.elem is None:
+            raise KeyError(localname)
+        return attribname
+
     def _getelem(
         self, localname: str
     ) -> "None | lxml.etree.Element | list[lxml.etree.Element]":
         """Return an element or tuple of elements given a localname."""
-        childdef = self.typedef.get_childdef_from_localname(localname)
-        if childdef is None:
-            raise KeyError(localname)
+        childdef = self._getchilddef(localname)
         if childdef.repeat:
             if self.elem is None:
                 return tuple()
@@ -205,11 +217,8 @@ class ElementWrapper(collections.abc.MutableMapping):
 
     def __getitem__(self, localname: str):
         if localname.startswith("@"):
-            attribname = self.typedef.get_attribute_from_localname(
-                localname.removeprefix("@")
-            )
-            if attribname is None or self.elem is None:
-                raise KeyError(localname)
+            attribname = self._getattribname(localname)
+            assert self.elem is not None  # for mypy.  Case handled by _getattribname()
             return self.elem.get(attribname)
 
         elem = self._getelem(localname)
@@ -260,17 +269,11 @@ class ElementWrapper(collections.abc.MutableMapping):
             self.wrapped_parent[lxml.etree.QName(elemtag).localname] = self.elem
 
         if localname.startswith("@"):
-            attribname = self.typedef.get_attribute_from_localname(
-                localname.removeprefix("@")
-            )
-            if attribname is None:
-                raise KeyError(localname)
+            attribname = self._getattribname(localname)
             self.elem.set(attribname, str(value))
             return
 
-        childdef = self.typedef.get_childdef_from_localname(localname)
-        if childdef is None:
-            raise KeyError(localname)
+        childdef = self._getchilddef(localname)
 
         transcoder = self.xsdhelper.get_transcoder(childdef.typename, childdef.tag)
 
@@ -319,9 +322,7 @@ class ElementWrapper(collections.abc.MutableMapping):
         Any
             the new transcoded or wrapped subelement
         """
-        childdef = self.typedef.get_childdef_from_localname(localname)
-        if childdef is None:
-            raise KeyError(localname)
+        childdef = self._getchilddef(localname)
 
         setval = val if val is not None else {}
 
@@ -335,16 +336,10 @@ class ElementWrapper(collections.abc.MutableMapping):
 
     def __delitem__(self, localname):
         if localname.startswith("@"):
-            attribname = self.typedef.get_attribute_from_localname(
-                localname.removeprefix("@")
-            )
-            if attribname is None:
-                raise KeyError(localname)
+            attribname = self._getattribname(localname)
             del self.elem.attrib[attribname]
         else:
-            childdef = self.typedef.get_childdef_from_localname(localname)
-            if childdef is None:
-                raise KeyError(localname)
+            childdef = self._getchilddef(localname)
             for subelem in self.elem.findall(childdef.tag):
                 self.elem.remove(subelem)
 
@@ -372,6 +367,12 @@ class ElementWrapper(collections.abc.MutableMapping):
         return len(self._keys())
 
     def __contains__(self, localname):
+        # Make sure localname is valid
+        if localname.startswith("@"):
+            _ = self._getattribname(localname)
+        else:
+            _ = self._getchilddef(localname)
+
         return localname in self._keys()
 
     def to_dict(self) -> dict: