+++ tests/core/xmlhelp/test_core.py
@@ -25,6 +25,9 @@ def test_elementwrapper():
     with pytest.raises(KeyError, match="foo"):
         del wrapped_siddroot["foo"]
 
+    with pytest.raises(KeyError, match="foo"):
+        "foo" in wrapped_siddroot
+
     # Attribute KeyErrors
     with pytest.raises(KeyError, match="@fooattr"):
         wrapped_siddroot["@fooattr"] = "doesn't exist"
@@ -34,6 +37,9 @@ def test_elementwrapper():
 
     with pytest.raises(KeyError, match="@fooattr"):
         del wrapped_siddroot["@fooattr"]
+
+    with pytest.raises(KeyError, match="@fooattr"):
+        "@fooattr" in wrapped_siddroot
 
     # Add descendant of repeatable
     wrapped_siddroot["ProductProcessing"].add("ProcessingModule")["ModuleName"] = (